Move organisation resolvers into dedicated modules
